+++ lib/ruler_picker_lib.dart
@@ -8,5 +8,10 @@ import 'package:flutter/services.dart';
 
 part 'ruler_picker/ruler_picker.dart';
 part 'ruler_picker/element/ruler_vertical_line.dart';
+part 'ruler_picker/element/ruler_aligned_long_vertical_line.dart';
+part 'ruler_picker/element/ruler_aligned_short_vertical_line.dart';
+part 'ruler_picker/element/ruler_short_vertical_line.dart';
+part 'ruler_picker/element/ruler_picked_short_vertical_line.dart';
+part 'ruler_picker/element/ruler_long_vertical_line.dart';
+part 'ruler_picker/element/ruler_picked_long_vertical_line.dart';
 
-// export 'ruler_picker/ruler_picker.dart' show MoonRulerPicker;
